backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
xmltodict
pydantic
rich
aiofiles
unidecode
python-multipart
itsdangerous
sendgrid

# --- Pacotes de IA ---
google-genai
openai>=1.0.0

# --- Pacotes de Auth e Banco de Dados ---
sqlalchemy
passlib[argon2]
python-jose
